Push before another attempt at creating a TCP Client/Server without using the Application Pattern.
This code contains all the code I used to test TCP Client/Server Application that was being created for this platform.

I am about to experiment with some code that will remove the Application superclass from the code due to so many problems occuring.

While NS-3's patterns are great when the simulator is running on its own, but the patterns are causing too many issues when run with Mosaik.