Upgrading OS and OpenDSS compatibility
Made the simulation run on Ubuntu 24.04, upgraded opendss as the old one is not available anymore.
Still has a few more bugs to fix like mosaik stepping more than max advance internally.
Reached out to mosaik for a fix, in the mean time will work on a temporary fix in a future commit.